+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add `debug` option to `AdvancedBuilder.expand`.
 - Fix
   - Parser can detect R-S conflict like: `` exp: `a b | a b c` `` where `b` is a terminator.
+  - Parser won't print `[user resolved RR]` when it's actually an R-S conflict.
 
 ## v0.6.0