+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix
   - Parser can detect R-S conflict like: `` exp: `a b | a b c` `` where `b` is a terminator.
   - Parser won't print `[user resolved RR]` when it's actually an R-S conflict.
+- More tests.
 
 ## v0.6.0