+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Parser
   - Add `debug` option to `AdvancedBuilder.expand`.
+  - Auto resolve R-S conflicts when `AdvancedBuilder.expand`.
 - Fix
   - Parser can detect R-S conflict like: `` exp: `a b | a b c` `` where `b` is a terminator.
   - Parser won't print `[user resolved RR]` when it's actually an R-S conflict.